Move lesson and day enum to new lesson package under person module
