+++ src/test/java/tuteez/logic/parser/ParserUtilTest.java
@@ -18,7 +18,7 @@ import tuteez.model.person.Address;
 import tuteez.model.person.Email;
 import tuteez.model.person.Name;
 import tuteez.model.person.Phone;
-import tuteez.model.tag.Lesson;
+import tuteez.model.person.lesson.Lesson;
 import tuteez.model.tag.Tag;
 
 public class ParserUtilTest {
